Update firestoreCommands to create FirestoreController. Created FirestoreController that is an interface to relevant firestore commands. Created embedCommands to create EmbedController. Created EmbedController that is an interface to relevant chat and embed functionality.
